django_backend/requirements.txt
django
djangorestframework
django-cors-headers
markdown
